Fixed: Upscaling operator with content negotiation
This commit contains a fix for content negotiation with IIIF. It moves the setting of `iiif_version` outside of the info.json handler so that it can be used for parsing image requests. This fixes a bug where using the upscaling operator (^) without having IIIF_VERSION specified in the configuration would cause the request to fail, since the handler assumed that it should be parsing it as a IIIF v2 request.