Code cleanup and refactoring
OK, I agree about keeping the region array as ratios. If I recall, the reason I had it like it was to avoid possible rounding errors, so we'll need to properly test that it now works as it should. I also simplified things so that session->view->setViewLeft() etc is only called once at the end of the section. Plus I eliminated some type casting warnings.