IIPServer: Pass Through Mode – Fix Colour Interpretation Issue
This commit fixes the colourspace interpretation issue, present since
pass-though was introduced last year; the colorspace interpretation
issue manifests in particular for Adobe style JPEGs that are missing
their APP14 marker segment which is used to denote the colorspace being
used so that decoders can correctly interpret and decode the
colourspace information in the image to correctly render the colors.

When the APP14 marker is missing, but required based on the format of
the JPEG, many decoders, particularly ones that follow colour
management best practices are susceptible to rendering the colour
information incorrectly, and often this results in the images being
rendered with a magenta/pink and green hue!

While pass-through functionality significantly improves throughput
speed, by avoiding the decompression/recompression of tiles, it can
result in necessary metadata being left out of the tile, leading to
rendering issues.

Tiles along the sides and edges of an image, are usually not affected
by the colourspace rendering issue because they are incomplete tiles,
and thus need to go through the regular Compressor() pathway in code
which includes fully assembling the JPEG including all its metadata.

This issue was found last year and a fix was implemented at that time
in a local development branch, but other work commitments precluded
finalizing it for a pull request. With further improvements to
IIPServer including the addition of the new
`Compressor::injectMetadata()` method, we now have a suitable location
into which this fix can be applied.

I have updated the `JPEGCompressor` class’ `injectMetadata()` method
with the fix, which involves creating a `jpeg_decompress_struct` and a
`jpeg_create_decompress` step to copy the already encoded coefficients
from the provided `RawTile` into the output, and then adding in the
necessary metadata. The decompression step does not actually cause the
JPEG data to be decompressed, we simply copy the already compressed
data from the input to the output, which is incredibly fast, and
because we are constructing the output JPEG in a standards compliant
way all the necessary metadata is being carried over from source image
to destination image, including if needed, the APP14 Adobe marker.

The code while slightly longer (mostly due to code comments) should
also handle a range of other edge cases without needing special
handling as we are just relying on the libjpeg or libturbojpeg library
methods.

This fix has been compared against the previous implementation of the
`insertMetadata()` method and for JPEG tiles that were previously
affected by this issue, the updated version of this method resolves the
issue and tiles render correctly, including in Safari and Preview on
macOS as well as within other colour managed software such as Photoshop.